BFS benchmark with chunk reader
Trying to load the graph in a single ocall,
would not be a good idea since if the graph size is more,
there is a need to increase the stack size of the untrusted
application by using the ulimit bash command / limit tcsh
command which would a user action.

Chunk reader should be introduced in benchmarks to load the
graph / pages into the memory.

Implements: bfs