Openssl benchmark with chunk reader
Trying to load the huge file in a single ocall,
would not be a good idea since if the file size is more,
there is a need to increase the stack size of the untrusted
application by using the ulimit bash command / limit tcsh
command which would a user action.

Chunk reader should be introduced in benchmarks to load the
data into the memory in chunk of size 4096 Bytes.

Implements: openssl